// native/src/graphql/__generated__/graphql.ts
export type Maybe<T> = T | null;
// ...
export type Scalars = {
  ID: { input: string; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
  /** An ISO 8601-encoded date */
  ISO8601Date: { input: any; output: any; }
  /** An ISO 8601-encoded datetime */
  ISO8601DateTime: { input: any; output: any; }
  /** Represents untyped JSON */
  JSON: { input: any; output: any; }
  Upload: { input: any; output: any; }
};
// ...
export type Attachment = {
  __typename?: 'Attachment';
  blob: Blob;
  blobId: Scalars['ID']['output'];
  byteSize: Scalars['Int']['output'];
  contentType: Scalars['String']['output'];
  createdAt: Scalars['ISO8601DateTime']['output'];
  filename: Scalars['String']['output'];
  updatedAt: Scalars['ISO8601DateTime']['output'];
  url: Scalars['String']['output'];
};
// ...
export type Blob = {
  __typename?: 'Blob';
  byteSize: Scalars['Int']['output'];
  contentType: Scalars['String']['output'];
  createdAt: Scalars['ISO8601DateTime']['output'];
  filename: Scalars['String']['output'];
  metadata?: Maybe<Scalars['JSON']['output']>;
  updatedAt: Scalars['ISO8601DateTime']['output'];
  url: Scalars['String']['output'];
};
// ...
export type Composition = {
  __typename?: 'Composition';
  compositionLyric: CompositionLyric;
  compositionLyrics: CompositionLyricConnection;
  compositionRole: CompositionRole;
  compositionRoles: CompositionRoleConnection;
  id?: Maybe<Scalars['ID']['output']>;
  lyric: Lyric;
  lyrics: LyricConnection;
  recording: Recording;
  recordings: RecordingConnection;
  title?: Maybe<Scalars['String']['output']>;
};
// ...
export type CompositionConnection = {
  __typename?: 'CompositionConnection';
  /** A list of edges. */
  edges: Array<CompositionEdge>;
  /** A list of nodes. */
  nodes: Array<Composition>;
  /** Information to aid in pagination. */
  pageInfo: PageInfo;
};

/** An edge in a connection. */
export type CompositionEdge = {
  __typename?: 'CompositionEdge';
  /** A cursor for use in pagination. */
  cursor: Scalars['String']['output'];
  /** The item at the end of the edge. */
  node?: Maybe<Composition>;
};

export type CompositionLyric = {
  __typename?: 'CompositionLyric';
  composition: Composition;
  compositionId: Scalars['ID']['output'];
  createdAt: Scalars['ISO8601DateTime']['output'];
  id: Scalars['ID']['output'];
  lyric: Lyric;
  lyricId: Scalars['ID']['output'];
  updatedAt: Scalars['ISO8601DateTime']['output'];
};

/** The connection type for CompositionLyric. */
export type CompositionLyricConnection = {
  __typename?: 'CompositionLyricConnection';
  /** A list of edges. */
  edges: Array<CompositionLyricEdge>;
  /** A list of nodes. */
  nodes: Array<CompositionLyric>;
  /** Information to aid in pagination. */
  pageInfo: PageInfo;
};

/** An edge in a connection. */
export type CompositionLyricEdge = {
  __typename?: 'CompositionLyricEdge';
  /** A cursor for use in pagination. */
  cursor: Scalars['String']['output'];
  /** The item at the end of the edge. */
  node?: Maybe<CompositionLyric>;
};

export type CompositionRole = {
  __typename?: 'CompositionRole';
  composition: Composition;
  compositionId: Scalars['ID']['output'];
  id?: Maybe<Scalars['ID']['output']>;
  person: Person;
  personId: Scalars['ID']['output'];
  role: CompositionRoleRole;
};

/** The connection type for CompositionRole. */
export type CompositionRoleConnection = {
  __typename?: 'CompositionRoleConnection';
  /** A list of edges. */
  edges: Array<CompositionRoleEdge>;
  /** A list of nodes. */
  nodes: Array<CompositionRole>;
  /** Information to aid in pagination. */
  pageInfo: PageInfo;
};
// ...
export type Genre = {
  __typename?: 'Genre';
  createdAt: Scalars['ISO8601DateTime']['output'];
  id: Scalars['ID']['output'];
  name: Scalars['String']['output'];
  recording: Recording;
  recordings: RecordingConnection;
  updatedAt: Scalars['ISO8601DateTime']['output'];
};


export type GenreRecordingArgs = {
  id: Scalars['ID']['input'];
};


export type GenreRecordingsArgs = {
  after?: InputMaybe<Scalars['String']['input']>;
  before?: InputMaybe<Scalars['String']['input']>;
  first?: InputMaybe<Scalars['Int']['input']>;
  last?: InputMaybe<Scalars['Int']['input']>;
};

/** The connection type for Genre. */
export type GenreConnection = {
  __typename?: 'GenreConnection';
  /** A list of edges. */
  edges: Array<GenreEdge>;
  /** A list of nodes. */
  nodes: Array<Genre>;
  /** Information to aid in pagination. */
  pageInfo: PageInfo;
};

/** An edge in a connection. */
export type GenreEdge = {
  __typename?: 'GenreEdge';
  /** A cursor for use in pagination. */
  cursor: Scalars['String']['output'];
  /** The item at the end of the edge. */
  node?: Maybe<Genre>;
};
// ...
export type Lyric = {
  __typename?: 'Lyric';
  composition: Composition;
  compositionId: Scalars['ID']['output'];
  createdAt?: Maybe<Scalars['ISO8601Date']['output']>;
  id?: Maybe<Scalars['ID']['output']>;
  language: Language;
  languageId: Scalars['ID']['output'];
  text: Scalars['String']['output'];
  updatedAt?: Maybe<Scalars['ISO8601Date']['output']>;
};

/** The connection type for Lyric. */
export type LyricConnection = {
  __typename?: 'LyricConnection';
  /** A list of edges. */
  edges: Array<LyricEdge>;
  /** A list of nodes. */
  nodes: Array<Lyric>;
  /** Information to aid in pagination. */
  pageInfo: PageInfo;
};
// ...
export type Orchestra = {
  __typename?: 'Orchestra';
  composition: Composition;
  compositions: CompositionConnection;
  genre: Genre;
  genres: GenreConnection;
  id?: Maybe<Scalars['ID']['output']>;
  image?: Maybe<Attachment>;
  name: Scalars['String']['output'];
  orchestraPeriod: OrchestraPeriod;
  orchestraPeriods: OrchestraPeriodConnection;
  orchestraPosition: OrchestraPosition;
  orchestraPositions: OrchestraPositionConnection;
  orchestraRole: OrchestraRole;
  orchestraRoles: OrchestraRoleConnection;
  recording: Recording;
  recordings: RecordingConnection;
  singer: Singer;
  singers: SingerConnection;
  slug?: Maybe<Scalars['String']['output']>;
  sortName?: Maybe<Scalars['String']['output']>;
};
// ...
export type OrchestraConnection = {
  __typename?: 'OrchestraConnection';
  /** A list of edges. */
  edges: Array<OrchestraEdge>;
  /** A list of nodes. */
  nodes: Array<Orchestra>;
  /** Information to aid in pagination. */
  pageInfo: PageInfo;
};

/** An edge in a connection. */
export type OrchestraEdge = {
  __typename?: 'OrchestraEdge';
  /** A cursor for use in pagination. */
  cursor: Scalars['String']['output'];
  /** The item at the end of the edge. */
  node?: Maybe<Orchestra>;
};

export type OrchestraPeriod = {
  __typename?: 'OrchestraPeriod';
  description?: Maybe<Scalars['String']['output']>;
  endDate?: Maybe<Scalars['ISO8601Date']['output']>;
  id?: Maybe<Scalars['ID']['output']>;
  name?: Maybe<Scalars['String']['output']>;
  orchestra: Orchestra;
  orchestraId: Scalars['ID']['output'];
  startDate?: Maybe<Scalars['ISO8601Date']['output']>;
};

/** The connection type for OrchestraPeriod. */
export type OrchestraPeriodConnection = {
  __typename?: 'OrchestraPeriodConnection';
  /** A list of edges. */
  edges: Array<OrchestraPeriodEdge>;
  /** A list of nodes. */
  nodes: Array<OrchestraPeriod>;
  /** Information to aid in pagination. */
  pageInfo: PageInfo;
};

/** An edge in a connection. */
export type OrchestraPeriodEdge = {
  __typename?: 'OrchestraPeriodEdge';
  /** A cursor for use in pagination. */
  cursor: Scalars['String']['output'];
  /** The item at the end of the edge. */
  node?: Maybe<OrchestraPeriod>;
};

export type OrchestraPosition = {
  __typename?: 'OrchestraPosition';
  createdAt?: Maybe<Scalars['ISO8601Date']['output']>;
  endDate?: Maybe<Scalars['ISO8601Date']['output']>;
  id?: Maybe<Scalars['ID']['output']>;
  orchestra: Orchestra;
  orchestraId: Scalars['ID']['output'];
  orchestraRole: OrchestraRole;
  orchestraRoleId: Scalars['ID']['output'];
  person: Person;
  personId: Scalars['ID']['output'];
  principal?: Maybe<Scalars['Boolean']['output']>;
  startDate?: Maybe<Scalars['ISO8601Date']['output']>;
  updatedAt?: Maybe<Scalars['ISO8601Date']['output']>;
};

/** The connection type for OrchestraPosition. */
export type OrchestraPositionConnection = {
  __typename?: 'OrchestraPositionConnection';
  /** A list of edges. */
  edges: Array<OrchestraPositionEdge>;
  /** A list of nodes. */
  nodes: Array<OrchestraPosition>;
  /** Information to aid in pagination. */
  pageInfo: PageInfo;
};

/** An edge in a connection. */
export type OrchestraPositionEdge = {
  __typename?: 'OrchestraPositionEdge';
  /** A cursor for use in pagination. */
  cursor: Scalars['String']['output'];
  /** The item at the end of the edge. */
  node?: Maybe<OrchestraPosition>;
};

export type OrchestraRole = {
  __typename?: 'OrchestraRole';
  id?: Maybe<Scalars['ID']['output']>;
  name?: Maybe<Scalars['String']['output']>;
  orchestra: Orchestra;
  orchestraPosition: OrchestraPosition;
  orchestraPositions: OrchestraPositionConnection;
  orchestras: OrchestraConnection;
};
// ...
export type OrchestraRoleConnection = {
  __typename?: 'OrchestraRoleConnection';
  /** A list of edges. */
  edges: Array<OrchestraRoleEdge>;
  /** A list of nodes. */
  nodes: Array<OrchestraRole>;
  /** Information to aid in pagination. */
  pageInfo: PageInfo;
};
// ...
export type PageInfo = {
  __typename?: 'PageInfo';
  /** When paginating forwards, the cursor to continue. */
  endCursor?: Maybe<Scalars['String']['output']>;
  /** When paginating forwards, are there more items? */
  hasNextPage: Scalars['Boolean']['output'];
  /** When paginating backwards, are there more items? */
  hasPreviousPage: Scalars['Boolean']['output'];
  /** When paginating backwards, the cursor to continue. */
  startCursor?: Maybe<Scalars['String']['output']>;
};
// ...
export type Recording = {
  __typename?: 'Recording';
  audioVariant: AudioVariant;
  audioVariants: AudioVariantConnection;
  composition?: Maybe<Composition>;
  compositionId?: Maybe<Scalars['ID']['output']>;
  digitalRemaster: DigitalRemaster;
  digitalRemasters: DigitalRemasterConnection;
  elRecodoSong?: Maybe<ElRecodoSong>;
  elRecodoSongId?: Maybe<Scalars['ID']['output']>;
  genre?: Maybe<Genre>;
  genreId?: Maybe<Scalars['ID']['output']>;
  id: Scalars['ID']['output'];
  like: Like;
  likedByCurrentUser?: Maybe<Scalars['Boolean']['output']>;
  likes: LikeConnection;
  lyric: Lyric;
  lyrics: LyricConnection;
  orchestra?: Maybe<Orchestra>;
  orchestraId?: Maybe<Scalars['ID']['output']>;
  playback: Playback;
  playbacks: PlaybackConnection;
  playbacksCount: Scalars['Int']['output'];
  playlistItem: PlaylistItem;
  playlistItems: PlaylistItemConnection;
  recordLabel?: Maybe<RecordLabel>;
  recordLabelId?: Maybe<Scalars['ID']['output']>;
  recordedDate?: Maybe<Scalars['ISO8601Date']['output']>;
  recordingSinger: RecordingSinger;
  recordingSingers: RecordingSingerConnection;
  recordingType: RecordingTypeEnum;
  share: Share;
  shares: ShareConnection;
  singer: Singer;
  singers: SingerConnection;
  slug: Scalars['String']['output'];
  tag: Tag;
  tagging: Tagging;
  taggings: TaggingConnection;
  tags: TagConnection;
  tanda: Tanda;
  tandaRecording: TandaRecording;
  tandaRecordings: TandaRecordingConnection;
  tandas: TandaConnection;
  timePeriod?: Maybe<TimePeriod>;
  timePeriodId?: Maybe<Scalars['ID']['output']>;
  waveform: Waveform;
  waveforms: WaveformConnection;
  year?: Maybe<Scalars['Int']['output']>;
};
// ...
export type RecordingConnection = {
  __typename?: 'RecordingConnection';
  /** A list of edges. */
  edges: Array<RecordingEdge>;
  /** A list of nodes. */
  nodes: Array<Recording>;
  /** Information to aid in pagination. */
  pageInfo: PageInfo;
};

/** An edge in a connection. */
export type RecordingEdge = {
  __typename?: 'RecordingEdge';
  /** A cursor for use in pagination. */
  cursor: Scalars['String']['output'];
  /** The item at the end of the edge. */
  node?: Maybe<Recording>;
};
// ...
export enum RecordingTypeEnum {
  Live = 'live',
  Studio = 'studio'
}
// ...
export type Singer = {
  __typename?: 'Singer';
  bio?: Maybe<Scalars['String']['output']>;
  birthDate?: Maybe<Scalars['ISO8601Date']['output']>;
  composition: Composition;
  compositionRole: CompositionRole;
  compositionRoles: CompositionRoleConnection;
  compositions: CompositionConnection;
  createdAt?: Maybe<Scalars['ISO8601Date']['output']>;
  deathDate?: Maybe<Scalars['ISO8601Date']['output']>;
  id?: Maybe<Scalars['ID']['output']>;
  image?: Maybe<Attachment>;
  name?: Maybe<Scalars['String']['output']>;
  orchestra: Orchestra;
  orchestraRole: OrchestraRole;
  orchestraRoles: OrchestraRoleConnection;
  orchestras: OrchestraConnection;
  recording: Recording;
  recordingSinger: RecordingSinger;
  recordingSingers: RecordingSingerConnection;
  recordings: RecordingConnection;
  slug?: Maybe<Scalars['String']['output']>;
  sortName?: Maybe<Scalars['String']['output']>;
  updatedAt?: Maybe<Scalars['ISO8601Date']['output']>;
};
// ...
export type SingerConnection = {
  __typename?: 'SingerConnection';
  /** A list of edges. */
  edges: Array<SingerEdge>;
  /** A list of nodes. */
  nodes: Array<Singer>;
  /** Information to aid in pagination. */
  pageInfo: PageInfo;
};